Fix opening IEC61937 8ch 192kHz on some Amazon devices
On Amazon Fire stick 4k Max 2023 codename kara, we can't open IEC61937
8ch 192kHz for TrueHD/DTS:X passthrough because of some stupidly wrong java check.
It turns out, Amazon made it pretty easy for us to override this check,
even though it's likely completely illegal APIs.
Please note that this bug is originally an AOSP bug that the vast
majority of platforms have, but Amazon made it easy to workaround.