edge to edge: get txt/gfx subs layout right even with floating player
- avoids round edges for subs
- takes video surface resize into consideration

Note of the design choice: to respect gfx subs on the top/bottom of the bitmap aligned with top/bottom of the screen, rescale the gfx sub to match the video surface. It has the side effect of not having the entire sub visible in portrait when filling the screen (but who uses this mode?).