Solves subtitles not being downloaded on nvidia shield external storage with two levels deep directory structure
Note that on the shield (and not on mibox gen3) when you are using a extfat or ntfs usb drive, Android reports that first level directories (/storage/XXX or /storage/XXX/serie) are writeable (`canWrite=true`) but not on deeper levels (e.g. /storage/XXX/serie/season1 where `canWrite=false`).
Note that with `canWrite=true` writing to the file generates no error and thus does not trigger a fallback mechanism.
To work around this issue, try to write something in the file to check if it is writeable and fallback to nova subtitle cache directory if file is not created.

See https://github.com/nova-video-player/aos-AVP/issues/1464