BUGFIX: Stretch to Screen Width NOT respecting CUTOUT, also cutting sides off  Cinema Sources on Phones.
* Stretch to Width uses Display Width (Minus Cutouts if Applicable) / AR for calulation now, not cropH.
* Always use same calculation.

IF (DCAR < AR) logic is wrong, explanation as follows:
Cinema Source at 2.39:1 will be used as Example.
My phone is 2340x1080-34 for hole cutout, DCAR of 2.135185185185185
DCAR is less than AR, so we cut width CropW in landscape on phone.

This caused the Cutout bug, and the overcrop on some sources...

DVD Source is 720x576@1.4 PAR, Effective A/R of 1.75 (for reference, not related to bug!)